Make consul-rust async.
Consul-rust currently uses reqwest::blocking::Client to issue HTTP
requests to Consul. As a result, it cannot be called from an
async runtime.

From the reqwest documentation:
"Conversely, the functionality in reqwest::blocking must not be
executed within an async runtime, or it will panic when attempting
to block."

This change makes the consul-rust library use reqwest::Client and
converts all calls to async calls. Tests are also reworked to be
called in an async runtime.